+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/FilterBar.test.tsx
@@ -29,7 +29,8 @@ import { TimeFilterPlugin, SelectFilterPlugin } from 'src/filters/components';
 import { DATE_FILTER_TEST_KEY } from 'src/explore/components/controls/DateFilterControl';
 import fetchMock from 'fetch-mock';
 import { waitFor } from '@testing-library/react';
-import FilterBar, { FILTER_BAR_TEST_ID } from '.';
+import { FILTER_BAR_TEST_ID } from './utils';
+import { VerticalFilterBar } from '.';
 import { FILTERS_CONFIG_MODAL_TEST_ID } from '../FiltersConfigModal/FiltersConfigModal';
 
 jest.useFakeTimers();
@@ -216,12 +217,15 @@ describe('FilterBar', () => {
   });
 
   const renderWrapper = (props = closedBarProps, state?: object) =>
-    render(<FilterBar {...props} width={280} height={400} offset={0} />, {
-      initialState: state,
-      useDnd: true,
-      useRedux: true,
-      useRouter: true,
-    });
+    render(
+      <VerticalFilterBar {...props} width={280} height={400} offset={0} />,
+      {
+        initialState: state,
+        useDnd: true,
+        useRedux: true,
+        useRouter: true,
+      },
+    );
 
   it('should render', () => {
     const { container } = renderWrapper();
